refactor: extract PeerAddr
