+++ src/services/cluster/manager.rs
@@ -2,7 +2,8 @@ use super::actors::actor::ClusterActor;
 use crate::services::cluster::command::ClusterCommand;
 use crate::services::cluster::inbound::stream::InboundStream;
 use crate::services::cluster::outbound::stream::OutboundStream;
-use crate::services::cluster::peer::identifier::{PeerAddrs, PeerIdentifier};
+use crate::services::cluster::peer::address::PeerAddrs;
+use crate::services::cluster::peer::identifier::PeerIdentifier;
 use crate::services::cluster::peer::kind::PeerKind;
 use crate::services::cluster::replication::replication::{Replication, IS_MASTER_MODE};
 use crate::services::statefuls::cache::manager::CacheManager;
