+++ src/services/cluster/peer/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod peer;
 pub(crate) mod connected_types;
 pub(crate) mod kind;
 pub(crate) mod identifier;
+pub(crate) mod address;
